requirements.txt
apache-airflow
requests
minio
pyspark==3.3.0